Move modules into package and add CLI entry
